Moved `wwctl service` to `wwctl system`
